Address review: shared captureSpec bounds, fullPage capture-time cap, capture-derived fullPage dims
The captureSpec bounds and strict parse move to capture-spec.ts so every
surface that accepts a spec off the wire enforces them identically: the
realm-server's POST /_screenshot-card delegates to the shared parse, and
the prerender server's /prerender-screenshot route — its own HTTP
surface, which previously cast the spec through unvalidated — now 400s
out-of-bounds and unknown fields by name before any render work.

fullPage's extent (the document's scroll size) is unknowable at parse
time, so the capture path enforces the same physical-pixel cap itself:
a document whose scroll size × the effective scale exceeds the Chromium
texture cap is refused by name instead of captured. Within the cap, the
reported fullPage dimensions now come from the captured PNG's IHDR
(divided back to CSS px by the scale in effect), so the response
describes exactly the image the caller received rather than a separate
scroll-size measurement taken before the capture.

Tests: prerender-server route 400s for oversize and unknown-field specs;
browser tests for the fullPage cap rejection at 3x and a within-cap
fullPage whose reported dims equal the PNG's. All verified against a
live prerender stack.
