+++ packages/realm-server/handlers/handle-screenshot-card.ts
@@ -7,13 +7,13 @@ import {
   findLiveInstanceGeneration,
   findMediaCacheEntry,
   isCaptureFormat,
+  parseScreenshotCaptureSpec,
   screenshotURLFor,
   touchMediaCacheEntryOnHit,
   type CaptureSpec,
   type DBAdapter,
   type MediaCacheEntry,
   type MediaCacheEntryKey,
-  type ScreenshotCaptureSpec,
   type ScreenshotPrerenderResponse,
 } from '@cardstack/runtime-common';
 import RealmPermissionChecker from '@cardstack/runtime-common/realm-permission-checker';
@@ -98,16 +98,19 @@ interface CaptureResult {
  * }
  * ```
  *
- * `captureSpec` is optional; every field within it is optional. It is validated
- * here (viewport ≤ 4096×16384, deviceScaleFactor ≤ 3, clip bounded by the same
- * caps as the viewport and within the viewport when one is given, physical
- * pixels per edge ≤ the Chromium texture cap, `fullPage` and `clip` mutually
- * exclusive) so the worker/prerenderer downstream can treat it as trusted. The
- * parse is strict per the capture-spec contract (`capture-spec.ts`): a field
- * the engine cannot honor is refused by name — never ignored — and
- * default-valued fields (`fullPage: false`, `deviceScaleFactor: 1`) are elided
- * so equal capture intents classify identically. Invalid specs return a 400
- * naming the offending field.
+ * `captureSpec` is optional; every field within it is optional. It is
+ * validated by the shared strict parse in `capture-spec.ts` (viewport ≤
+ * 4096×16384, deviceScaleFactor ≤ 3, clip bounded by the same caps as the
+ * viewport and within the viewport when one is given, physical pixels per
+ * edge ≤ the Chromium texture cap, `fullPage` and `clip` mutually exclusive
+ * — the prerender server's screenshot route runs the identical parse), so
+ * the worker downstream can treat it as trusted. The parse is strict: a
+ * field the engine cannot honor is refused by name — never ignored — and
+ * default-valued fields (`fullPage: false`, `deviceScaleFactor: 1`) are
+ * elided so equal capture intents classify identically. Invalid specs
+ * return a 400 naming the offending field. The one bound no request-time
+ * parse can enforce — a fullPage capture's document extent — is checked
+ * against the same physical-pixel cap at capture time.
  * A spec with overrides is capture-only: the ledger's canonical capture
  * identity cannot represent it yet (the GET DSL reserves those params), so it
  * always renders fresh and returns raw bytes — no ledger fast path, no
@@ -116,240 +119,9 @@ interface CaptureResult {
  * The `runAs` user is derived from the authenticated JWT.
  */
 
-// Chromium caps a single texture at 16384px; a viewport wider than 4096px is
-// well past any real card layout and mostly a way to force a huge capture.
-export const SCREENSHOT_MAX_VIEWPORT_WIDTH = 4096;
-export const SCREENSHOT_MAX_VIEWPORT_HEIGHT = 16384;
-// A 3× scale already covers retina/hi-dpi; higher just multiplies pixel cost.
-export const SCREENSHOT_MAX_DEVICE_SCALE_FACTOR = 3;
-// The Chromium single-texture cap the viewport bounds are derived from,
-// enforced on *physical* pixels: CSS dimension × deviceScaleFactor. The CSS
-// caps alone would admit e.g. a 16384-tall viewport at 3× (~49k physical px).
-export const SCREENSHOT_MAX_PHYSICAL_EDGE_PX = 16384;
-
-// Result of validating the request's `captureSpec` attribute. On success
-// `captureSpec` is the normalized spec — null when the attribute was absent
-// or carried no overrides (default-valued fields are elided), so `null`
-// exactly means "the canonical capture"; on failure `error` names the
-// offending field for a 400.
-type CaptureSpecParse =
-  | { captureSpec: ScreenshotCaptureSpec | null; error?: undefined }
-  | { captureSpec?: undefined; error: string };
-
-const CAPTURE_SPEC_FIELDS = new Set([
-  'viewport',
-  'deviceScaleFactor',
-  'fullPage',
-  'clip',
-]);
-const CAPTURE_SPEC_VIEWPORT_FIELDS = new Set(['width', 'height']);
-const CAPTURE_SPEC_CLIP_FIELDS = new Set(['x', 'y', 'width', 'height']);
-
-function isPositiveInteger(value: unknown): value is number {
-  return typeof value === 'number' && Number.isInteger(value) && value > 0;
-}
-
-function isPlainObject(value: unknown): value is Record<string, unknown> {
-  return typeof value === 'object' && value !== null && !Array.isArray(value);
-}
-
-function parseCaptureSpec(raw: unknown): CaptureSpecParse {
-  if (raw === undefined || raw === null) {
-    return { captureSpec: null };
-  }
-  if (!isPlainObject(raw)) {
-    return { error: 'captureSpec must be an object' };
-  }
-
-  // Strict per the capture-spec contract: an unrecognized field is refused by
-  // name rather than dropped — silently dropping a typo'd `fullpage: true`
-  // would classify the request as canonical and serve (or persist) the wrong
-  // image.
-  for (let key of Object.keys(raw)) {
-    if (!CAPTURE_SPEC_FIELDS.has(key)) {
-      return { error: `captureSpec.${key} is not a supported field` };
-    }
-  }
-
-  let spec: ScreenshotCaptureSpec = {};
-
-  if (raw.viewport !== undefined) {
-    let viewport = raw.viewport;
-    if (!isPlainObject(viewport)) {
-      return {
-        error:
-          'captureSpec.viewport must have positive integer width and height',
-      };
-    }
-    for (let key of Object.keys(viewport)) {
-      if (!CAPTURE_SPEC_VIEWPORT_FIELDS.has(key)) {
-        return {
-          error: `captureSpec.viewport.${key} is not a supported field`,
-        };
-      }
-    }
-    if (
-      !isPositiveInteger(viewport.width) ||
-      !isPositiveInteger(viewport.height)
-    ) {
-      return {
-        error:
-          'captureSpec.viewport must have positive integer width and height',
-      };
-    }
-    if (viewport.width > SCREENSHOT_MAX_VIEWPORT_WIDTH) {
-      return {
-        error: `captureSpec.viewport.width must be <= ${SCREENSHOT_MAX_VIEWPORT_WIDTH}`,
-      };
-    }
-    if (viewport.height > SCREENSHOT_MAX_VIEWPORT_HEIGHT) {
-      return {
-        error: `captureSpec.viewport.height must be <= ${SCREENSHOT_MAX_VIEWPORT_HEIGHT}`,
-      };
-    }
-    spec.viewport = { width: viewport.width, height: viewport.height };
-  }
-
-  if (raw.deviceScaleFactor !== undefined) {
-    let scale = raw.deviceScaleFactor;
-    if (typeof scale !== 'number' || !Number.isFinite(scale) || scale <= 0) {
-      return {
-        error: 'captureSpec.deviceScaleFactor must be a positive number',
-      };
-    }
-    if (scale > SCREENSHOT_MAX_DEVICE_SCALE_FACTOR) {
-      return {
-        error: `captureSpec.deviceScaleFactor must be <= ${SCREENSHOT_MAX_DEVICE_SCALE_FACTOR}`,
-      };
-    }
-    // 1 is the engine default: elide it so `{ deviceScaleFactor: 1 }` means
-    // the same capture as no spec at all and keeps the canonical fast path.
-    if (scale !== 1) {
-      spec.deviceScaleFactor = scale;
-    }
-  }
-
-  if (raw.fullPage !== undefined) {
-    if (typeof raw.fullPage !== 'boolean') {
-      return { error: 'captureSpec.fullPage must be a boolean' };
-    }
-    // false is the engine default: elide it (see deviceScaleFactor above).
-    if (raw.fullPage) {
-      spec.fullPage = true;
-    }
-  }
-
-  if (raw.clip !== undefined) {
-    let clip = raw.clip;
-    if (!isPlainObject(clip)) {
-      return {
-        error:
-          'captureSpec.clip must have non-negative x/y and positive integer width/height',
-      };
-    }
-    for (let key of Object.keys(clip)) {
-      if (!CAPTURE_SPEC_CLIP_FIELDS.has(key)) {
-        return { error: `captureSpec.clip.${key} is not a supported field` };
-      }
-    }
-    if (
-      typeof clip.x !== 'number' ||
-      typeof clip.y !== 'number' ||
-      !Number.isFinite(clip.x) ||
-      !Number.isFinite(clip.y) ||
-      clip.x < 0 ||
-      clip.y < 0 ||
-      !isPositiveInteger(clip.width) ||
-      !isPositiveInteger(clip.height)
-    ) {
-      return {
-        error:
-          'captureSpec.clip must have non-negative x/y and positive integer width/height',
-      };
-    }
-    // The clip's extent is bounded by the same caps as the viewport whether
-    // or not one was sent: Puppeteer captures beyond the viewport by default
-    // (`captureBeyondViewport`), so an unbounded clip would be a way around
-    // the viewport cost caps.
-    if (clip.x + clip.width > SCREENSHOT_MAX_VIEWPORT_WIDTH) {
-      return {
-        error: `captureSpec.clip x + width must be <= ${SCREENSHOT_MAX_VIEWPORT_WIDTH}`,
-      };
-    }
-    if (clip.y + clip.height > SCREENSHOT_MAX_VIEWPORT_HEIGHT) {
-      return {
-        error: `captureSpec.clip y + height must be <= ${SCREENSHOT_MAX_VIEWPORT_HEIGHT}`,
-      };
-    }
-    spec.clip = {
-      x: clip.x,
-      y: clip.y,
-      width: clip.width,
-      height: clip.height,
-    };
-  }
-
-  if (spec.fullPage && spec.clip) {
-    return {
-      error: 'captureSpec cannot set both fullPage and clip',
-    };
-  }
-
-  // Tighter containment when the caller declared a viewport: the layout was
-  // requested at that size, so a clip past its edge is caller error — the
-  // region would be unstyled overflow or blank area past the document edge.
-  if (spec.clip && spec.viewport) {
-    if (spec.clip.x + spec.clip.width > spec.viewport.width) {
-      return {
-        error: 'captureSpec.clip exceeds captureSpec.viewport.width',
-      };
-    }
-    if (spec.clip.y + spec.clip.height > spec.viewport.height) {
-      return {
-        error: 'captureSpec.clip exceeds captureSpec.viewport.height',
-      };
-    }
-  }
-
-  // The CSS caps compose with the scale factor: what Chromium renders is
-  // physical pixels, and each capture edge must stay under the texture cap.
-  let effectiveScale = spec.deviceScaleFactor ?? 1;
-  if (spec.viewport) {
-    if (
-      spec.viewport.width * effectiveScale >
-      SCREENSHOT_MAX_PHYSICAL_EDGE_PX
-    ) {
-      return {
-        error: `captureSpec.viewport.width × deviceScaleFactor must be <= ${SCREENSHOT_MAX_PHYSICAL_EDGE_PX} physical pixels`,
-      };
-    }
-    if (
-      spec.viewport.height * effectiveScale >
-      SCREENSHOT_MAX_PHYSICAL_EDGE_PX
-    ) {
-      return {
-        error: `captureSpec.viewport.height × deviceScaleFactor must be <= ${SCREENSHOT_MAX_PHYSICAL_EDGE_PX} physical pixels`,
-      };
-    }
-  }
-  if (spec.clip) {
-    if (spec.clip.width * effectiveScale > SCREENSHOT_MAX_PHYSICAL_EDGE_PX) {
-      return {
-        error: `captureSpec.clip.width × deviceScaleFactor must be <= ${SCREENSHOT_MAX_PHYSICAL_EDGE_PX} physical pixels`,
-      };
-    }
-    if (spec.clip.height * effectiveScale > SCREENSHOT_MAX_PHYSICAL_EDGE_PX) {
-      return {
-        error: `captureSpec.clip.height × deviceScaleFactor must be <= ${SCREENSHOT_MAX_PHYSICAL_EDGE_PX} physical pixels`,
-      };
-    }
-  }
-
-  // A spec whose every field matched an engine default normalizes to null:
-  // it means the canonical capture, and null is what the rest of the handler
-  // (and the job args) key that classification on.
-  return { captureSpec: Object.keys(spec).length > 0 ? spec : null };
-}
+// The captureSpec bounds and strict parse live in `capture-spec.ts`
+// (runtime-common) so this handler and the prerender server's screenshot
+// route validate identically; see the constants and rules there.
 
 export default function handleScreenshotCard({
   dbAdapter,
@@ -420,7 +192,7 @@ export default function handleScreenshotCard({
     let sourceURL = normalizedCardId.replace(/\.json$/, '');
     let instanceLocalPath = sourceURL.slice(normalizedRealmURL.length);
 
-    let captureSpecParse = parseCaptureSpec(attrs.captureSpec);
+    let captureSpecParse = parseScreenshotCaptureSpec(attrs.captureSpec);
     if (captureSpecParse.error) {
       return sendResponseForBadRequest(ctxt, captureSpecParse.error);
     }
