+++ packages/runtime-common/capture-spec.ts
@@ -35,6 +35,254 @@ export function hasCaptureSpecOverrides(
   return !!captureSpec && Object.keys(captureSpec).length > 0;
 }
 
+// ---------------------------------------------------------------------------
+// ScreenshotCaptureSpec bounds + strict parse — one enforcement point for
+// every surface that accepts a spec off the wire (the realm-server's POST
+// /_screenshot-card body and the prerender server's /prerender-screenshot
+// route), and the home of the caps the capture path itself enforces for the
+// extents only it can know (a fullPage capture's document size).
+// ---------------------------------------------------------------------------
+
+// Chromium caps a single texture at 16384px; a viewport wider than 4096px is
+// well past any real card layout and mostly a way to force a huge capture.
+export const SCREENSHOT_MAX_VIEWPORT_WIDTH = 4096;
+export const SCREENSHOT_MAX_VIEWPORT_HEIGHT = 16384;
+// A 3× scale already covers retina/hi-dpi; higher just multiplies pixel cost.
+export const SCREENSHOT_MAX_DEVICE_SCALE_FACTOR = 3;
+// The Chromium single-texture cap the viewport bounds are derived from,
+// enforced on *physical* pixels: CSS dimension × deviceScaleFactor. The CSS
+// caps alone would admit e.g. a 16384-tall viewport at 3× (~49k physical px).
+// A fullPage capture's extent (the document's scroll size) is unknowable at
+// parse time, so the capture path checks it against this cap itself.
+export const SCREENSHOT_MAX_PHYSICAL_EDGE_PX = 16384;
+
+// Result of validating a raw `captureSpec` value. On success `captureSpec`
+// is the normalized spec — null when the value was absent or carried no
+// overrides (default-valued fields are elided), so `null` exactly means
+// "the canonical capture"; on failure `error` names the offending field.
+export type ScreenshotCaptureSpecParse =
+  | { captureSpec: ScreenshotCaptureSpec | null; error?: undefined }
+  | { captureSpec?: undefined; error: string };
+
+const CAPTURE_SPEC_FIELDS = new Set([
+  'viewport',
+  'deviceScaleFactor',
+  'fullPage',
+  'clip',
+]);
+const CAPTURE_SPEC_VIEWPORT_FIELDS = new Set(['width', 'height']);
+const CAPTURE_SPEC_CLIP_FIELDS = new Set(['x', 'y', 'width', 'height']);
+
+function isPositiveInteger(value: unknown): value is number {
+  return typeof value === 'number' && Number.isInteger(value) && value > 0;
+}
+
+function isPlainObject(value: unknown): value is Record<string, unknown> {
+  return typeof value === 'object' && value !== null && !Array.isArray(value);
+}
+
+export function parseScreenshotCaptureSpec(
+  raw: unknown,
+): ScreenshotCaptureSpecParse {
+  if (raw === undefined || raw === null) {
+    return { captureSpec: null };
+  }
+  if (!isPlainObject(raw)) {
+    return { error: 'captureSpec must be an object' };
+  }
+
+  // Strict per the capture-spec contract: an unrecognized field is refused by
+  // name rather than dropped — silently dropping a typo'd `fullpage: true`
+  // would classify the request as canonical and serve (or persist) the wrong
+  // image.
+  for (let key of Object.keys(raw)) {
+    if (!CAPTURE_SPEC_FIELDS.has(key)) {
+      return { error: `captureSpec.${key} is not a supported field` };
+    }
+  }
+
+  let spec: ScreenshotCaptureSpec = {};
+
+  if (raw.viewport !== undefined) {
+    let viewport = raw.viewport;
+    if (!isPlainObject(viewport)) {
+      return {
+        error:
+          'captureSpec.viewport must have positive integer width and height',
+      };
+    }
+    for (let key of Object.keys(viewport)) {
+      if (!CAPTURE_SPEC_VIEWPORT_FIELDS.has(key)) {
+        return {
+          error: `captureSpec.viewport.${key} is not a supported field`,
+        };
+      }
+    }
+    if (
+      !isPositiveInteger(viewport.width) ||
+      !isPositiveInteger(viewport.height)
+    ) {
+      return {
+        error:
+          'captureSpec.viewport must have positive integer width and height',
+      };
+    }
+    if (viewport.width > SCREENSHOT_MAX_VIEWPORT_WIDTH) {
+      return {
+        error: `captureSpec.viewport.width must be <= ${SCREENSHOT_MAX_VIEWPORT_WIDTH}`,
+      };
+    }
+    if (viewport.height > SCREENSHOT_MAX_VIEWPORT_HEIGHT) {
+      return {
+        error: `captureSpec.viewport.height must be <= ${SCREENSHOT_MAX_VIEWPORT_HEIGHT}`,
+      };
+    }
+    spec.viewport = { width: viewport.width, height: viewport.height };
+  }
+
+  if (raw.deviceScaleFactor !== undefined) {
+    let scale = raw.deviceScaleFactor;
+    if (typeof scale !== 'number' || !Number.isFinite(scale) || scale <= 0) {
+      return {
+        error: 'captureSpec.deviceScaleFactor must be a positive number',
+      };
+    }
+    if (scale > SCREENSHOT_MAX_DEVICE_SCALE_FACTOR) {
+      return {
+        error: `captureSpec.deviceScaleFactor must be <= ${SCREENSHOT_MAX_DEVICE_SCALE_FACTOR}`,
+      };
+    }
+    // 1 is the engine default: elide it so `{ deviceScaleFactor: 1 }` means
+    // the same capture as no spec at all and keeps the canonical fast path.
+    if (scale !== 1) {
+      spec.deviceScaleFactor = scale;
+    }
+  }
+
+  if (raw.fullPage !== undefined) {
+    if (typeof raw.fullPage !== 'boolean') {
+      return { error: 'captureSpec.fullPage must be a boolean' };
+    }
+    // false is the engine default: elide it (see deviceScaleFactor above).
+    if (raw.fullPage) {
+      spec.fullPage = true;
+    }
+  }
+
+  if (raw.clip !== undefined) {
+    let clip = raw.clip;
+    if (!isPlainObject(clip)) {
+      return {
+        error:
+          'captureSpec.clip must have non-negative x/y and positive integer width/height',
+      };
+    }
+    for (let key of Object.keys(clip)) {
+      if (!CAPTURE_SPEC_CLIP_FIELDS.has(key)) {
+        return { error: `captureSpec.clip.${key} is not a supported field` };
+      }
+    }
+    if (
+      typeof clip.x !== 'number' ||
+      typeof clip.y !== 'number' ||
+      !Number.isFinite(clip.x) ||
+      !Number.isFinite(clip.y) ||
+      clip.x < 0 ||
+      clip.y < 0 ||
+      !isPositiveInteger(clip.width) ||
+      !isPositiveInteger(clip.height)
+    ) {
+      return {
+        error:
+          'captureSpec.clip must have non-negative x/y and positive integer width/height',
+      };
+    }
+    // The clip's extent is bounded by the same caps as the viewport whether
+    // or not one was sent: Puppeteer captures beyond the viewport by default
+    // (`captureBeyondViewport`), so an unbounded clip would be a way around
+    // the viewport cost caps.
+    if (clip.x + clip.width > SCREENSHOT_MAX_VIEWPORT_WIDTH) {
+      return {
+        error: `captureSpec.clip x + width must be <= ${SCREENSHOT_MAX_VIEWPORT_WIDTH}`,
+      };
+    }
+    if (clip.y + clip.height > SCREENSHOT_MAX_VIEWPORT_HEIGHT) {
+      return {
+        error: `captureSpec.clip y + height must be <= ${SCREENSHOT_MAX_VIEWPORT_HEIGHT}`,
+      };
+    }
+    spec.clip = {
+      x: clip.x,
+      y: clip.y,
+      width: clip.width,
+      height: clip.height,
+    };
+  }
+
+  if (spec.fullPage && spec.clip) {
+    return {
+      error: 'captureSpec cannot set both fullPage and clip',
+    };
+  }
+
+  // Tighter containment when the caller declared a viewport: the layout was
+  // requested at that size, so a clip past its edge is caller error — the
+  // region would be unstyled overflow or blank area past the document edge.
+  if (spec.clip && spec.viewport) {
+    if (spec.clip.x + spec.clip.width > spec.viewport.width) {
+      return {
+        error: 'captureSpec.clip exceeds captureSpec.viewport.width',
+      };
+    }
+    if (spec.clip.y + spec.clip.height > spec.viewport.height) {
+      return {
+        error: 'captureSpec.clip exceeds captureSpec.viewport.height',
+      };
+    }
+  }
+
+  // The CSS caps compose with the scale factor: what Chromium renders is
+  // physical pixels, and each capture edge must stay under the texture cap.
+  // fullPage's extent (the document scroll size) isn't knowable here; the
+  // capture path enforces the same cap on it at capture time.
+  let effectiveScale = spec.deviceScaleFactor ?? 1;
+  if (spec.viewport) {
+    if (
+      spec.viewport.width * effectiveScale >
+      SCREENSHOT_MAX_PHYSICAL_EDGE_PX
+    ) {
+      return {
+        error: `captureSpec.viewport.width × deviceScaleFactor must be <= ${SCREENSHOT_MAX_PHYSICAL_EDGE_PX} physical pixels`,
+      };
+    }
+    if (
+      spec.viewport.height * effectiveScale >
+      SCREENSHOT_MAX_PHYSICAL_EDGE_PX
+    ) {
+      return {
+        error: `captureSpec.viewport.height × deviceScaleFactor must be <= ${SCREENSHOT_MAX_PHYSICAL_EDGE_PX} physical pixels`,
+      };
+    }
+  }
+  if (spec.clip) {
+    if (spec.clip.width * effectiveScale > SCREENSHOT_MAX_PHYSICAL_EDGE_PX) {
+      return {
+        error: `captureSpec.clip.width × deviceScaleFactor must be <= ${SCREENSHOT_MAX_PHYSICAL_EDGE_PX} physical pixels`,
+      };
+    }
+    if (spec.clip.height * effectiveScale > SCREENSHOT_MAX_PHYSICAL_EDGE_PX) {
+      return {
+        error: `captureSpec.clip.height × deviceScaleFactor must be <= ${SCREENSHOT_MAX_PHYSICAL_EDGE_PX} physical pixels`,
+      };
+    }
+  }
+
+  // A spec whose every field matched an engine default normalizes to null:
+  // it means the canonical capture, and null is what consumers key that
+  // classification on.
+  return { captureSpec: Object.keys(spec).length > 0 ? spec : null };
+}
+
 // The DSL's parameter surface grows with the capture engine; these names are
 // reserved for the engine capabilities the project's URL grammar assigns
 // them, and refused (never ignored) while the engine lacks them.
